remove extra code
