+++ components/ordhook-core/src/core/meta_protocols/brc20/mod.rs
@@ -2,7 +2,6 @@ use chainhook_sdk::types::BitcoinNetwork;
 
 pub mod brc20_pg;
 pub mod cache;
-pub mod db;
 pub mod parser;
 pub mod test_utils;
 pub mod verifier;
